Setup XO VS Code integration
The XO VSCode doesn't work with yarn PNP out of the box. Someone found
a solution by createing a patched node runtime that can be used to run
XO.
https://github.com/xojs/vscode-linter-xo/issues/71
https://github.com/flying-sheep/react-color-scheme-switch/pull/3
https://github.com/yarnpkg/berry/issues/4045

I added the XO plugin to the recommended VSCode extensions and added
a workspace config to automatically enable it for the project.