Create a CLI that can create a basic React PDF
I want to use TypeScript. That's why it is necessary to build the
project with `tsc`.
https://www.digitalocean.com/community/tutorials/typescript-new-project
https://www.digitalocean.com/community/tutorials/setting-up-a-node-project-with-typescript

I re-ran the yarn Editor SDKs setup to enable TypeScript support.

I had to disable the XO rule `import/extensions` because it required
that `./layout/cv` is imported with `.tsx` and TypeScript does not allow
that.

I added the cli as the `bin` of the project so that it will be available
via `npx`.

I used parameter passing of `run-s` (from `npm-run-all`) to be able to
pass the arguments of `yarn cli` into `dist/cli.js`.
https://github.com/mysticatea/npm-run-all/blob/master/docs/run-s.md#argument-placeholders

I used React PDF to be able to style the CV via React that is familiar
to me.
https://react-pdf.org/

I had to add `@types/node` to the TypeScript config to be able to use
Node buildin modules.
https://stackoverflow.com/a/45446470