Add `init` command to the CLI
The `init` command creates a git repository that hosts the data
configuration files so that they can be stored independently in a
(private) git repo.

I added fs-extra to be able to use their `copy` function to easily copy
the init-tempalte files to the new data project.
https://github.com/jprichardson/node-fs-extra

The `init-template` folder has to be copied to the dist folder to be
able to use it from the published npm package (will be published in the
future)

I added a `--data-project` option to the `create` command to be able to
run the `create` command from any folder and not just from within a
data project. It also helps with running `yarn cli create` from within
this repository on a data project somewhere else. Example to use the
template data to create a CV:
```
yarn cli create -d src/init-template
```

I moved the schema into the `init-template` folder because the schemas
should be available in the data project to make it easier to edit the
configuration files.

I used the `--initial-branch` option of `git init` to set the main
branch to `main` because the default is still `master`.
https://twitter.com/mislav/status/1270388510684598272

The `-C` option of `git` makes it possible to run a git command in a
different folder.
https://stackoverflow.com/a/20115753