Show contact & personal data in the CV header
It seemed more fitting to define the data in contact & personal instead
of a header file that is tightly coupled to where the information is
shown instead of what the information represents.

I used a fictitious German telephone number.
https://en.m.wikipedia.org/wiki/Fictitious_telephone_number#Germany

I formatted the telephone number according to the E.123 international
standard.
https://en.wikipedia.org/wiki/National_conventions_for_writing_telephone_numbers#Germany

I used the folded block style indicator so that it's possible to write
the summary without having to write very long lines.
> If instead you want them to be replaced by spaces, use the folded
> style, indicated by a right angle bracket (>). (To get a newline
> using the folded style, leave a blank line by putting two newlines in.
> Lines with extra indentation are also not folded.)
> https://yaml-multiline.info/

The logic of only applying styles if there was a previous sibling is
necessary because the styles of ReactPDF don't support the CSS concept
of :first-child or :last-child.