Add an education section to the CV
I split the information items into `title` and `content` to be able
to style the title differently (e.g. bold).

I decided to allow a max of 7 items in an information section to avoid
that the lists are too big and hard to follow. I decided to use a max of
3 items to describe an education to avoid that the education section is
too long (focus should be on the work experience).

The education layout is pretty similar to the work-experience layout. I
did not create reusable components because the actual styling of the CV
will be done in a future branch.