Add Link & SubTitleLink components
I decided to not highlight link so that they look a bit more subtle and
refined (especially when printed).

There is no concept of inheritance and `currentColor` is not implemented
in ReactPDF. That's why a `Link` inside a `SubTitle` won't have the
color of the `SubTitle`. That makes a `SubTitleLink` component
necessary.
https://github.com/diegomura/react-pdf/issues/3
https://github.com/diegomura/react-pdf/issues/1708