Add an image to the header
The image is optional because some companies might require a CV without
an image.

I added a smaller lineHeight to the name so the job title is closer to
it and it looks more like a unit. I also made the name strong because it
was always planned like this.