Refactor horizontal list styling (header)
The previous implementation required extra logic as to when to apply the
list style that depended on the fact if the previous item existed. The
HorizontalList ensures that styling by only allowing existing items to
be passed in which makes the styling decisions easier. It also makes the
ContactAndLocation JSX easier to follow.