Refactor SubTitle to contain logic
Both sub titles contain the same logic (link when website is available,
add the location if it is available, add the date information). I moved
the logic into the SubTitle so it is not repeated and the JSX of
education and work-experience is easier to follow.

I decided to use a HorizontalList to make it a bit easier to see the
date when scanning the information.