Make font.base configurable
The fonts no longer need to be copied to the dist folder manually since
they are inside the init-template folder now.

registerFonts had to be refactored to load the font files from a given
project path to be able to load the font files that are defined in the
data project.

I chose to make getFontSource a curried function to be able to create a
specialized function for the path and family combination that can be
reused for multiple styles (less duplication).